correctly display completed achievements
