+++ app/src/main/java/io/github/viabachelora23michaelkutaibakasper/bprapp/MainActivity.kt
@@ -13,7 +13,7 @@ import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.core.content.ContextCompat
-import io.github.viabachelora23michaelkutaibakasper.bprapp.ui.theme.BPRAppTheme
+import io.github.viabachelora23michaelkutaibakasper.bprapp.ui.theme.AppTheme
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.asStateFlow
 
@@ -72,5 +72,5 @@ class MainActivity : ComponentActivity() {
 @Preview(showBackground = true)
 @Composable
 fun GreetingPreview() {
-    BPRAppTheme {}
+    AppTheme {}
 }
