+++ app/src/main/java/io/github/viabachelora23michaelkutaibakasper/bprapp/ui/screens/events/map/MapView.kt
@@ -83,7 +83,7 @@ import io.github.viabachelora23michaelkutaibakasper.bprapp.data.domain.User
 import io.github.viabachelora23michaelkutaibakasper.bprapp.ui.navigation.BottomNavigationScreens
 import io.github.viabachelora23michaelkutaibakasper.bprapp.ui.navigation.CreateEventScreens
 import io.github.viabachelora23michaelkutaibakasper.bprapp.ui.navigation.navigateTo
-import io.github.viabachelora23michaelkutaibakasper.bprapp.ui.theme.BPRAppTheme
+import io.github.viabachelora23michaelkutaibakasper.bprapp.ui.theme.AppTheme
 import io.github.viabachelora23michaelkutaibakasper.bprapp.util.DisplayFormattedTime
 import io.github.viabachelora23michaelkutaibakasper.bprapp.util.localDateTimeToUTCLocalDateTime
 import java.time.LocalDateTime
@@ -449,7 +449,7 @@ fun EventListItem(event: MinimalEvent, navController: NavController) {
 @Preview(showBackground = true)
 @Composable
 fun GreetingPreview() {
-    BPRAppTheme {
+    AppTheme {
         EventListItem(
             event = MinimalEvent(
                 title = "Title",
