+++ app/src/main/java/io/github/viabachelora23michaelkutaibakasper/bprapp/ui/theme/Color.kt
@@ -2,10 +2,67 @@ package io.github.viabachelora23michaelkutaibakasper.bprapp.ui.theme
 
 import androidx.compose.ui.graphics.Color
 
-val Purple80 = Color(0xFFD0BCFF)
-val PurpleGrey80 = Color(0xFFCCC2DC)
-val Pink80 = Color(0xFFEFB8C8)
+val md_theme_light_primary = Color(0xFF0056D2)
+val md_theme_light_onPrimary = Color(0xFFFFFFFF)
+val md_theme_light_primaryContainer = Color(0xFFDAE2FF)
+val md_theme_light_onPrimaryContainer = Color(0xFF001848)
+val md_theme_light_secondary = Color(0xFF805600)
+val md_theme_light_onSecondary = Color(0xFFFFFFFF)
+val md_theme_light_secondaryContainer = Color(0xFFFFDDAF)
+val md_theme_light_onSecondaryContainer = Color(0xFF281800)
+val md_theme_light_tertiary = Color(0xFF8A5100)
+val md_theme_light_onTertiary = Color(0xFFFFFFFF)
+val md_theme_light_tertiaryContainer = Color(0xFFFFDCBE)
+val md_theme_light_onTertiaryContainer = Color(0xFF2C1600)
+val md_theme_light_error = Color(0xFFBA1A1A)
+val md_theme_light_errorContainer = Color(0xFFFFDAD6)
+val md_theme_light_onError = Color(0xFFFFFFFF)
+val md_theme_light_onErrorContainer = Color(0xFF410002)
+val md_theme_light_background = Color(0xFFFFFBFF)
+val md_theme_light_onBackground = Color(0xFF2A1800)
+val md_theme_light_surface = Color(0xFFFFFBFF)
+val md_theme_light_onSurface = Color(0xFF2A1800)
+val md_theme_light_surfaceVariant = Color(0xFFE1E2EC)
+val md_theme_light_onSurfaceVariant = Color(0xFF45464F)
+val md_theme_light_outline = Color(0xFF757780)
+val md_theme_light_inverseOnSurface = Color(0xFFFFEEDE)
+val md_theme_light_inverseSurface = Color(0xFF462A00)
+val md_theme_light_inversePrimary = Color(0xFFB2C5FF)
+val md_theme_light_shadow = Color(0xFF000000)
+val md_theme_light_surfaceTint = Color(0xFF0056D2)
+val md_theme_light_outlineVariant = Color(0xFFC5C6D0)
+val md_theme_light_scrim = Color(0xFF000000)
 
-val Purple40 = Color(0xFF6650a4)
-val PurpleGrey40 = Color(0xFF625b71)
-val Pink40 = Color(0xFF7D5260)
+val md_theme_dark_primary = Color(0xFFB2C5FF)
+val md_theme_dark_onPrimary = Color(0xFF002B73)
+val md_theme_dark_primaryContainer = Color(0xFF0040A1)
+val md_theme_dark_onPrimaryContainer = Color(0xFFDAE2FF)
+val md_theme_dark_secondary = Color(0xFFFFBA44)
+val md_theme_dark_onSecondary = Color(0xFF442C00)
+val md_theme_dark_secondaryContainer = Color(0xFF614000)
+val md_theme_dark_onSecondaryContainer = Color(0xFFFFDDAF)
+val md_theme_dark_tertiary = Color(0xFFFFB86F)
+val md_theme_dark_onTertiary = Color(0xFF4A2800)
+val md_theme_dark_tertiaryContainer = Color(0xFF693C00)
+val md_theme_dark_onTertiaryContainer = Color(0xFFFFDCBE)
+val md_theme_dark_error = Color(0xFFFFB4AB)
+val md_theme_dark_errorContainer = Color(0xFF93000A)
+val md_theme_dark_onError = Color(0xFF690005)
+val md_theme_dark_onErrorContainer = Color(0xFFFFDAD6)
+val md_theme_dark_background = Color(0xFF2A1800)
+val md_theme_dark_onBackground = Color(0xFFFFDDB6)
+val md_theme_dark_surface = Color(0xFF2A1800)
+val md_theme_dark_onSurface = Color(0xFFFFDDB6)
+val md_theme_dark_surfaceVariant = Color(0xFF45464F)
+val md_theme_dark_onSurfaceVariant = Color(0xFFC5C6D0)
+val md_theme_dark_outline = Color(0xFF8F909A)
+val md_theme_dark_inverseOnSurface = Color(0xFF2A1800)
+val md_theme_dark_inverseSurface = Color(0xFFFFDDB6)
+val md_theme_dark_inversePrimary = Color(0xFF0056D2)
+val md_theme_dark_shadow = Color(0xFF000000)
+val md_theme_dark_surfaceTint = Color(0xFFB2C5FF)
+val md_theme_dark_outlineVariant = Color(0xFF45464F)
+val md_theme_dark_scrim = Color(0xFF000000)
+
+
+val seed = Color(0xFF105CDA)
